benthos pooled sql update
